backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
pytest
httpx
